Implement stub loading
When starting the dedicated server from NorthstarLauncher, always
default to loading the d3d11 and gfsdk stubs from `bin/x64_dedi`. If
the stubs fail to load or `-nostubs` is specified, the launcher reverts
to the old behaviour of using the GPU, or WARP if `-softwared3d11` is
specified.

updates #79