Update to match recent LatencyFleX changes
Main entry point for LatencyFlex was changed from latencyflex_wine.dll to latencyflex_layer.dll to prepare library for eventual windows release/to be more generic. This updates our implementation to match, while still supporting the previous libraries for existing users.

Currently in the process of testing, not ready to merge.