Update Origin -> Auth Provider
With the public deprecation of Origin, our messaging should reflect this change. This changes some text strings to better reflect the current state. Also adds the `-noAuthProviderStartup` argument, though this can be removed/replace `-noOriginStartup` argument if wanted.

I'm not sure of the process name of EA App (is it just EADesktop?), so im keeping the old Origin.exe check in for now, though this should probably be removed in the future.